+++ CHANGELOG.rst
@@ -1,3 +1,14 @@
+Unreleased
+----------
+
+Date
+  2024-02-02
+
+Changed
+ * Refresh device states when MQTT connection is resumed
+ * Try to recreate MQTT clients after connection is lost. This is an additional
+   measure on top of AWS SDK trying to resume connection.
+
 Version 0.6
 -----------